third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-border-repeat-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-border-repeat with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-border-repeat">
<meta name="assert" content="mask-border-repeat supports only the grammar '[ stretch | repeat | round | space ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-border-repeat", "auto");
test_invalid_value("mask-border-repeat", "0");
test_invalid_value("mask-border-repeat", "repeat round space");
</script>
</body>
</html>
